+++ src/components/emissionFactor/new/Form.tsx
@@ -1,8 +1,5 @@
 'use client'
 import Form from '@/components/base/Form'
-import LoadingButton from '@/components/base/LoadingButton'
-import { FormSelect } from '@/components/form/Select'
-import { FormTextField } from '@/components/form/TextField'
 import { createEmissionFactorCommand } from '@/services/serverFunctions/emissionFactor'
 import {
   CreateEmissionFactorCommand,
@@ -10,19 +7,15 @@ import {
   maxParts,
 } from '@/services/serverFunctions/emissionFactor.command'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { MenuItem } from '@mui/material'
-import { Unit } from '@prisma/client'
 import { useTranslations } from 'next-intl'
 import { useRouter } from 'next/navigation'
-import { FormEvent, useMemo, useState } from 'react'
+import { FormEvent, useState } from 'react'
 import { useForm } from 'react-hook-form'
-import DetailedGES from './DetailedGES'
-import Posts from './Posts'
+import EmissionFactorForm from '../Form/EmissionFactorForm'
 
 const NewEmissionFactorForm = () => {
   const router = useRouter()
   const t = useTranslations('emissionFactors.create')
-  const tUnit = useTranslations('units')
   const [error, setError] = useState('')
   const [hasParts, setHasParts] = useState(false)
   const [partsCount, setPartsCount] = useState(1)
@@ -57,44 +50,17 @@ const NewEmissionFactorForm = () => {
     })()
   }
 
-  const units = useMemo(() => Object.values(Unit).sort((a, b) => tUnit(a).localeCompare(tUnit(b))), [tUnit])
   return (
     <Form onSubmit={onSubmit}>
-      <FormTextField
-        data-testid="new-emission-name"
-        control={form.control}
-        translation={t}
-        name="name"
-        label={t('name')}
-      />
-      <FormTextField control={form.control} translation={t} name="attribute" label={t('attribute')} />
-      <FormTextField
-        data-testid="new-emission-source"
-        control={form.control}
-        translation={t}
-        name="source"
-        label={t('source')}
-      />
-      <FormSelect data-testid="new-emission-unit" control={form.control} translation={t} label={t('unit')} name="unit">
-        {units.map((unit) => (
-          <MenuItem key={unit} value={unit}>
-            {tUnit(unit)}
-          </MenuItem>
-        ))}
-      </FormSelect>
-      <DetailedGES
+      <EmissionFactorForm
         form={form}
+        error={error}
+        t={t}
         hasParts={hasParts}
         setHasParts={setHasParts}
         partsCount={partsCount}
         setPartsCount={setPartsCount}
       />
-      <Posts form={form} />
-      <FormTextField control={form.control} translation={t} name="comment" label={t('comment')} multiline rows={2} />
-      <LoadingButton type="submit" loading={form.formState.isSubmitting} data-testid="new-emission-create-button">
-        {t('create')}
-      </LoadingButton>
-      {error && <p>{error}</p>}
     </Form>
   )
 }
